add LineStyle option
