+++ Editor/Interfaces/ExportOptions.cs
@@ -5,19 +5,6 @@ using System.Collections.Generic;
 
 namespace asmdef2pu.Interfaces
 {
-    [Serializable]
-    internal enum DirectionStyle : byte
-    {
-        TopToBottom,
-        BottomToTop
-    }
-
-    [Serializable]
-    internal class StyleOptions
-    {
-        public DirectionStyle DirectionStyle = DirectionStyle.TopToBottom;
-    }
-
     /// <summary>
     /// Options for the target assembly of the generated graph
     /// </summary>
